Fix -Wsign-compare: use size_t for loop indices compared against size_t bounds

+++ src/xc_integrator/replicated/host/reference_replicated_xc_host_integrator_dd_psi_potential.hpp
@@ -70,7 +70,7 @@ void ReferenceReplicatedXCHostIntegrator<ValueType>::
 
   // Atom-specific data
   std::vector<double> radii(mol.size());
-  for (int i = 0; i < mol.size(); ++i) {
+  for (size_t i = 0; i < mol.size(); ++i) {
     radii[i] = uff_radius_103(mol[i].Z);
   }
